requirements.txt
requests>=2.28.0
python-chess>=1.999
